Refactor exception handling

+++ Api/Controllers/BooksController.cs
@@ -1,3 +1,4 @@
+using LibraryAPI.Core.Exceptions;
 using LibraryAPI.Core.IServices;
 using LibraryAPI.Core.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -13,11 +14,7 @@ namespace LibraryAPI.Api.Controllers
         [HttpGet]
         public IActionResult GetAllBooks()
         {
-            var books = _bookService.GetAllBooks();
-            if (books == null)
-            {
-                return StatusCode(500, new { message = $"An unexpected error occurred." });
-            }
+            var books = _bookService.GetAllBooks() ?? throw new NotFoundException("Book records not found.");
 
             return Ok(books);
         }
@@ -25,27 +22,18 @@ namespace LibraryAPI.Api.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetBook(int id)
         {
-            try
-            {
-                var book = await _bookService.GetBookDetails(id);
-                if (book == null)
-                {
-                    return NotFound();
-                }
+            var book = await _bookService.GetBookDetails(id) ?? throw new NotFoundException("Book record not found.");
 
-                return Ok(book);
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new { message = $"An unexpected error occurred: {ex.Message}" });
-            }
+            return Ok(book);
         }
 
         [HttpPost]
         public async Task<IActionResult> CreateBook([FromBody] Book book)
         {
             if (!ModelState.IsValid)
-                return BadRequest(ModelState);
+            {
+                throw new BadRequestException("Input parameter is not valid.");
+            }
 
             await _bookService.CreateBook(book);
             return CreatedAtAction(nameof(GetBook), new { id = book.Id }, book);
@@ -55,10 +43,14 @@ namespace LibraryAPI.Api.Controllers
         public async Task<IActionResult> UpdateBook(int id, [FromBody] Book book)
         {
             if (!ModelState.IsValid)
-                return BadRequest(ModelState);
+            {
+                throw new BadRequestException("Input parameter is not valid.");
+            }
 
             if (id != book.Id)
-                return BadRequest();
+            {
+                throw new BadRequestException("ID of book is different in body and parameter.");
+            }
 
             await _bookService.UpdateBook(book);
             return Ok();
@@ -67,6 +59,11 @@ namespace LibraryAPI.Api.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteBook(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                throw new BadRequestException("Input parameter is not valid.");
+            }
+
             await _bookService.DeleteBook(id);
             return Ok();
         }
